+++ tests/test_processing.py
@@ -1,6 +1,17 @@
 import pytest
 
-from src.processing import sort_by_date
+from src.processing import filter_by_state, sort_by_date
+
+
+def test_filter_by_state(
+    dict_list_state, return_dict_by_executed, return_dict_by_canceled, message_error_for_filter_by_state
+):
+    assert filter_by_state(dict_list_state, "CANCELED") == return_dict_by_canceled
+    assert filter_by_state(dict_list_state, "") == message_error_for_filter_by_state
+    assert filter_by_state(dict_list_state, "EXECUTED") == return_dict_by_executed
+    assert filter_by_state(dict_list_state, "executed") == message_error_for_filter_by_state
+    assert filter_by_state(dict_list_state, "canceled") == message_error_for_filter_by_state
+    assert filter_by_state(dict_list_state, "canceled") == message_error_for_filter_by_state
 
 
 def test_sort_by_date(dict_for_sort, result_sort_date):
